Dart 2.0 dev.39.0 Fixes related to convert.

+++ lib/src/element/bytes/ivr.dart
@@ -76,7 +76,7 @@ abstract class IvrElement<V> implements BDElement<V> {
 
   /// Returns the Value Field Length field.
   @override
-  int get vfLengthField => bytes.getUint16(_vfLengthOffset);
+  int get vfLengthField => bytes.getUint32(_vfLengthOffset);
 
   @override
   Bytes get vfBytesWithPadding =>
